Remove SingleVariable
