+++ src/FileFormats/NL/NLExpr.jl
@@ -296,8 +296,6 @@ end
 
 _NLExpr(x::MOI.VariableIndex) = _NLExpr(true, _NLTerm[], Dict(x => 1.0), 0.0)
 
-_NLExpr(x::MOI.SingleVariable) = _NLExpr(x.variable)
-
 function _add_or_set(dict, key, value)
     if haskey(dict, key)
         dict[key] += value
